+++ godot/addons/cyclops_level_builder/actions/action_export_as_godot_scene.gd
@@ -22,13 +22,13 @@
 # SOFTWARE.
 
 @tool
-class_name ActionExportAsGltf
+class_name ActionExportAsGodotScene
 extends CyclopsAction
 
-var wizard:ExporterGltfWizard = preload("res://addons/cyclops_level_builder/exporter/exporter_gltf_wizard.tscn").instantiate()
+var wizard:ExporterGodotSceneWizard = preload("res://addons/cyclops_level_builder/exporter/exporter_godot_scene_wizard.tscn").instantiate()
 
 func _init(plugin:CyclopsLevelBuilder, name:String = "", accellerator:Key = KEY_NONE):
-	super._init(plugin, "Export As Gltf...")
+	super._init(plugin, "Export As Godot Scene...")
 
 func _execute():
 	var base_control:Node = plugin.get_editor_interface().get_base_control()