+++ doc/change_log.md
@@ -10,3 +10,4 @@
 * Now generating tangent vectors for block meshes.
 * Restoring ability for UVs to remain in place when moving blocks.
 * Adding option to clear UV transform in material brush.
+* Adding tool to export current scene as Godot scene with only native Godot objects.